made the purpose of this application more specific
this project started as an effort to extract the bits we needed from the XML into our new article-json format. As the article-json format improved it was obvious the XML wouldn't be able to provide a complete set of data. the scope of this project has been reduced to just what the bot gives us and talking to Lax. It joins the EJP adaptor and soon, the CMS adaptor.