Fixing graceful shutdown
From
https://ci--alfred.elifesciences.org/job/test-elife-lax/110/consoleFull
```
      ID: bot-lax-adaptors-start
Function: cmd.run
    Name: start bot-lax-adaptors
  Result: True
 Comment: Command "start

 Started: 11:52:22.340042
Duration: 30129.05 ms
 Changes:
          ----------
          pid:
              3412
          retcode:
              0
          stderr:
          stdout:
              bot-lax-adaptors stop/waiting
```

The graceful shutdown times out after 30 seconds and the process gets
killed with `SIGKILL`.

The reason is if there are no messages (which is 99% of the time), the
inner while loop returns to polling without yielding a new value,
never exiting.

Therefore, passing the flag to the IncomingQueue allows it to check
whether it should stop, before starting any new polling. If polling
returned something, that is always processed before attempting a new
iteration.