Fix UTF-8 problems when downloading an XML file
Long story short, the XML in the S3 bucket are stored and served with `application/xml` as `Content-Type`, that doesn't specify an encoding. So `requests` stick tries some heuristics to guess the encoding but ultimately has undefined behavior.

Here we force these response to always be treated like they contain
UTF-8.