Generate IIIF links that use the IIIF CDNs
IIIF servers have a CDN now, so we can point there for caching on the
edge the tiles and resizes being generated.

When accessing the info.json files, we directly go to the IIIF server
since it's in the same availability zone and (maybe) data center, and we
cache the result locally.