Log error responses
The loggers set up on stderr and on adaptor.log ignore the extra parameter, hence error responses are not logged. By putting the responses inside the formatted message they become accessible.

Moreover, the message in http_download does not need to take into account the context of the file being an XML, as this is redundant with the message set in the `except` clause that catches it