Workaround to keep validating the kitchen sink
In this build we usually do not depend on external servers to contain test data, although we depend on production servers such as prod--iiif and the real, production of Glencoe.
However since we force the IIIF images to be existing, we would need an IIIF server where the bot has ingested the kitchen sink figure to be able to generate it. That, or to stub a dozen different calls to info.json files.

Therefore, a test setting to turn on selectively in the kicthen-sinks.sh script seems the lesser of three evils. Let's see if this makes the build green again